Restructure database directories to run in prod
Previously, we had specified relative directories for the paths to
tinypg's query SQL files and to postgres-migrate's DDL SQL files. That
worked fine in dev mode, where we were always running in the root of the
project and where we had the full source code available. However,
neither of those assumptions are true in production: we specify the
absolute path to the application entry (approximately,
`${DEPLOY_PATH}/lib/index.js`) which systemd calls without(?) setting
the working directory, and our build process excludes the src/
directory.

In fixing that, and after further thought, I realized that the previous
structure of our source tree was not ideal. The directory name `src/db/`
is too brief, and doesn't match the name of our "service"; likewise,
having `src/services/database.ts` live in a different directory - and
alongside files specifying business logic - was confusing and hard to
discover.

Rename the directory to `src/database/`, move the configuration code
into it, and tell the configuration code how to look for its SQL files
relative to itself.

As this is the first direct reference to the Node.js API, and we are
using version 12, also add in the v12 type definitions. We have a
transitive dependency on them already, and they were resolved to the v14
types, but nothing explicitly requires that version.

See also:
- https://nodejs.org/api/modules.html#modules_dirname
- https://stackoverflow.com/a/18283508/
- https://github.com/joeandaverde/tinypg#sql-files-over-embedded-strings
- https://github.com/ThomWright/postgres-migrations#api

Issue #20 Need to build and package notification-service
