+++ package.json
@@ -41,6 +41,7 @@
     "@tsconfig/node12": "^1.0.7",
     "@types/express": "^4.17.8",
     "@types/jest": "^26.0.13",
+    "@types/node": "^12.20.7",
     "@types/pg": "^7.14.10",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.1.1",
