+++ src/database/queries/health.sql
