+++ src/database/queries/insert-notification.sql
