+++ src/services/health.service.ts
@@ -1,4 +1,4 @@
-import { db } from './database';
+import { db } from '../database';
 
 enum HealthStatus {
   AVAILABLE = 'available',
