+++ src/services/notification.service.ts
@@ -1,4 +1,4 @@
-import { db } from './database';
+import { db } from '../database';
 
 export interface Notification {
   toUserId: number;
